Improvements to code that modifies TCL source
In change 9fafdcf, Eric made a small but significant change to the TCL code in gui.tcl that initializes the config file location:
- The CONFIG_FILE definition now uses $CONFIGDIR instead of $ROOTDIR
- On all platforms except Windows, $CONFIGDIR is initialized to $::env(HOME).

The Pascal scripting code in the installer previously changed $ROOTDIR to $::env(HOME) so the TCL code basically does what the installer already did, EXCEPT it doesn't do it for Windows.

The easy solution would be to add a little bit of code to remove the line in gui.tcl that changes the directory back to $ROOTDIR in Windows after setting it to $::env(HOME) in the first place. But we also have to deal with the possibility that the installer is doing an upgrade. So we still have to correctly replace the base filename of the installer (currently .flexprop.conf but previously .flexgui.conf).

I rewrote the Pascal scripting to be a bit more sophisticated so it can deal with this change and future changes, if necessary:
- Search and replace operations are now split up into separate routines that can locate strings in string arrays, and replace entire lines or partial lines as necessary
- Initialization of the config file name was separated in a way that's hopefully easier to read and maintain
- Modifications to the TCL code are now in a separate function that's hopefully also easier to read and maintain
- Added a missing function to initialize the config file name during uninstall so it gets correctly deleted