Significantly reduced installer complexity
Now that the TCL/TK code does the Right Thing (TM) and checks for the config file in $ROOTDIR first, but uses $HOME as default, the installer doesn't need to modify the application anymore to make it always look in the User Profile directory.

Previously the installer also modified the name of the config file if the application was being upgraded from a version before 4.5.0 (when it was still called flexgui). Instead of that, we now just rename the config file if an old file is found but no new file is present.

The installer checks for a config file in the program directory at install time, and deletes it if there is one. The only way this could happen is if someone previously unzipped the flexprop.zip file in the same location as where our installation lands. I considered moving such a file to the home directory but this should be a rare use case so I'm not going to bother adding code to do that.